+++ src/request/otel.rs
@@ -61,6 +61,13 @@ mod otel_crates {
     pub use tracing_opentelemetry_0_30_pkg as tracing_opentelemetry;
 }
 
+#[cfg(feature = "tracing_opentelemetry_0_30")]
+mod otel_crates {
+    pub use opentelemetry_0_30_pkg as opentelemetry;
+    pub use opentelemetry_sdk_0_30_pkg as opentelemetry_sdk;
+    pub use tracing_opentelemetry_0_31_pkg as tracing_opentelemetry;
+}
+
 use otel_crates::*;
 
 pub use opentelemetry::propagation::Injector;